chart_qa_extension/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
Pillow>=9.0.0
accelerate>=0.20.0
datasets>=2.14.0
